% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_alignRas.R
\name{mm_alignRas}
\alias{mm_alignRas}
\title{Create raster to align to afterwards}
\usage{
mm_alignRas(x, rN)
}
\arguments{
\item{x}{Raster that is used to define the extent}

\item{rN}{Desired resolution of the newly created raster}
}
\value{
New raster
}
\description{
Requires terra package.
Create raster with same or slightly larger extent than a given raster, but with a new desired resolution. Extent may be enlarged in all directions, so that the exact desired resolution fits in.
Output can be used for terra::resample()
}
\examples{
lidar = data("lidar.tif");
alignRas = mm_alignRas(x=lidar,rN=100);
x = terra::resample(lidar, alignRas, method="cubic", threads=TRUE);
}
