% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_bab.R
\name{mm_bab}
\alias{mm_bab}
\title{Create split buffer around a line (one polygon for right and left side respectively)}
\usage{
mm_bab(profile = pro$line, width = NA, s = stds$orographicSide)
}
\arguments{
\item{profile}{SpatVector. A Line, around which the buffer should be created.}

\item{width}{Numeric. Width of the buffer.}

\item{s}{Numeric. Either 1 (buffer for side A), 2 (side B), or 3 (both sides).}
}
\value{
SpatVector
}
\description{
Requires terra package
Takes SpatVector (profile line). Output is a SpatVector. This is used within mm_f(), to choose (profile-)orographic right of left side respectively.
Output is a SpatVector with two polygons, one on the left, and one on the right side of the given line.
}
\examples{
# create double sided buffer;
mm_bab(x = "notInPackage/testdata/profile.gpkg", width = 3000, side = 3);

}
