% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_metering.R
\name{mm_metering}
\alias{mm_metering}
\title{Create points along a line with attributes for metering}
\usage{
mm_metering(profile_line, spacing = 1000, label = T, labelUnit = "km")
}
\arguments{
\item{profile_line}{spatVector, line, along which points are placed.}

\item{spacing}{Distance between points in meters.}

\item{label}{logical. Standard = T. Should attributes for labelling = distance along the line, be set?}

\item{labelUnit}{Either "km" or "m". Standard = "km". Should label attributes be in kilometers or meters?}
}
\value{
spatVector
}
\description{
Requires sf and terra packages.
Creates points along a line with given spacing, starting at line distance = 0. The points are attributed by distance along the line. Output, spatVector, can be used for adding metering along a line in map plots.
Output can be used e.g. for terra::plot or plet
}
\examples{
profile = data("profile.gpkg");
metering = mm_metering(profile_line=profile,spacing=10000,label=T,labelUnit="km");
}
