% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_prepMap.R
\name{mm_prepMap}
\alias{mm_prepMap}
\title{Load, crop and rasterize polygon map}
\usage{
mm_prepMap(map, field, cropper, aligner, asVector = T)
}
\arguments{
\item{map}{character (path to) or spatVector of a polygon shapefile.}

\item{field}{Field of the map, that schould be kept.}

\item{cropper}{spatVector, Polygon to which the map should be cropped (i.e. profile_buffer).}

\item{aligner}{spatRaster to which the rasterized map should be aligned.}

\item{asVector}{logical. Should the output be a spatVector = T or a spatRaster = F.}
}
\value{
spatVector or spatRaster
}
\description{
Requires terra package.
Loads a polygon file or takes a spatVector of Polygons. Then clips it with a given polygon. Rasterizes map, aligned to given raster, keeping specified field. Output can be chosen as spatRaster or spatVector of points.
Output is a spatRast, or spatVector of points. Only specified fields are kept.
}
\examples{
profile_buffer = terra::buffer(x = terra::vect("notInPackage/testdata/profile.gpkg"), width = 3000, capstyle = "flat", joinstyle = "round");
ras = mm_prepRas(profile=profile_line, 3000, c("notInPackage/testdata/lidar.tif","notInPackage/testdata/lidarFilled.tif"), 50, makeSlope = T, makeShade = T);
map = mm_prpMap(map="notInPackage/testdata/terraceMap.gpkg",field="NAME_KURZ",cropper=profile_buffer,aligner=ras,asVector=T);
}
