% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_prepRas.R
\name{mm_prepRas}
\alias{mm_prepRas}
\title{Load, crop and align rasters}
\usage{
mm_prepRas(
  profile,
  searchRadius,
  raster,
  analysisReso,
  makeSlope = T,
  makeShade = T
)
}
\arguments{
\item{profile}{character (path to) or spatVector of a profile line.}

\item{searchRadius}{Radius around the profile to create a buffer to clip rasters with.}

\item{raster}{vector or list of character (path to) and or spatRaster.}

\item{analysisReso}{Resolution to which all rasters should be resampled.}

\item{makeSlope}{logical. Should a slope raster be added for the first provided raster?}

\item{makeShade}{logical. Should a hillshade be added for the first provided raster?}
}
\value{
spatRaster
}
\description{
Requires terra package.
Loads one or more rasters, that can be given as a vector or list of raster paths or spatRasters or mixed. The rasters are cropped and resampled to a desired analysis resolution and search radius around a given Profile. The profile can be given as path to, or spatVector.
Output is a spatRast, with each imported raster as a layer.
}
\examples{
ras = mm_prepRas(profile="profile.gpkg",searchRadius=3000,raster=c("data/lidar.tif","data/lidarFilled.tif"),analysisReso=100, makeSlope = T, makeShade = T);
}
